backend/requirements.txt
fastapi
uvicorn[standard]
weaviate-client
httpx
